# Optimise query_range by computing join signatures just once